% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_class.R
\docType{class}
\name{Api-class}
\alias{Api-class}
\alias{api_class}
\title{Api class to make appropriate request based on parameters}
\description{
Api class to make appropriate request based on parameters
}
\section{Fields}{

\describe{
\item{\code{api_type}}{character. The name of the api to call (see url_mapping)}

\item{\code{environment}}{environment on which to query API
Available are 'PRD', 'ABN', 'TEST', 'QA' and 'DEV'.}

\item{\code{export_format}}{character (default = "JSON"). The export's format
Available are CSV, XLSX, SDMX-ML and JSON}

\item{\code{parameters}}{character. Additional request parameters}

\item{\code{id}}{character. The identifier or id of the request's object}

\item{\code{language}}{character (default = "all"). Language of the response data.
Available are 'all', 'fr', 'de', 'it', 'en'}

\item{\code{version_format}}{numeric (default = 2.1). The export format's version
(2.0 or 2.1 when format is SDMX-ML)
(for 'codelist')}

\item{\code{api_url}}{character. The url to make the request to.}
}}


