% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_request.R
\name{csv_request}
\alias{csv_request}
\title{API query for CSV output}
\usage{
csv_request(url)
}
\arguments{
\item{url}{url to query}
}
\value{
dataframe response
}
\description{
API query for CSV output
}
