% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_call.R
\name{get_codelist}
\alias{get_codelist}
\title{Get a codelist based on an identifier}
\usage{
get_codelist(
  identifier,
  environment = "PRD",
  language = "all",
  export_format = "SDMX-ML",
  version_format = 2.1,
  annotations = FALSE
)
}
\arguments{
\item{identifier}{the codelist's identifier}

\item{environment}{environment on which to query API
Available are 'PRD', 'ABN', 'TEST', 'QA' and 'DEV'.}

\item{language}{string for language(s) to return.
Available are 'all', 'fr', 'de', 'it', 'en'.
If 'all' (default), all languages are returned.}

\item{export_format}{the export's format
Available are CSV, XLSX, SDMX-ML or JSON.}

\item{version_format}{the export format's version
(2.0 or 2.1 when format is SDMX-ML).}

\item{annotations}{flag to include annotations}
}
\value{
response based on the export format
}
\description{
Get a codelist based on an identifier
}
