% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_call.R
\name{get_nomenclature_multiple_levels}
\alias{get_nomenclature_multiple_levels}
\title{Get multiple levels of a nomenclature (from `level_from` to `level_to`)}
\usage{
get_nomenclature_multiple_levels(
  identifier,
  environment = "PRD",
  filters = list(),
  level_from = 1,
  level_to = 2,
  language = "fr",
  annotations = FALSE
)
}
\arguments{
\item{identifier}{nomenclature's identifier}

\item{environment}{environment on which to query API
Available are 'PRD', 'ABN', 'TEST', 'QA' and 'DEV'.}

\item{filters}{additionnal filters}

\item{level_from}{the 1st level to include}

\item{level_to}{the last level to include}

\item{language}{the language of the response data
Available are 'fr', 'de', 'it', 'en'.}

\item{annotations}{flag to include annotations}
}
\value{
dataframe columns
from `level_from` to `level_to` codes
}
\description{
Get multiple levels of a nomenclature (from `level_from` to `level_to`)
}
