% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_call.R
\name{get_nomenclature_one_level}
\alias{get_nomenclature_one_level}
\title{Get one level of a nomenclature}
\usage{
get_nomenclature_one_level(
  identifier,
  environment = "PRD",
  filters = list(),
  level_number = 1,
  language = "all",
  annotations = FALSE
)
}
\arguments{
\item{identifier}{nomenclature's identifier}

\item{environment}{environment on which to query API
Available are 'PRD', 'ABN', 'TEST', 'QA' and 'DEV'.}

\item{filters}{additionnal filters}

\item{level_number}{level to export}

\item{language}{the language of the response data
Available are 'fr', 'de', 'it', 'en'.}

\item{annotations}{flag to include annotations}
}
\value{
dataframe with 3 columns
(Code, Parent and Name in the selected language)
}
\description{
Get one level of a nomenclature
}
