% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_request.R
\name{json_request}
\alias{json_request}
\title{API query for JSON output}
\usage{
json_request(url)
}
\arguments{
\item{url}{url to query}
}
\value{
response: dataframe or list of dataframes
}
\description{
API query for JSON output
}
