% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_request.R
\name{list_to_string}
\alias{list_to_string}
\title{Transform a list into a string of parameters
list("a" = list("1"), "b" = list("2", "3"))
becomes "a=1&b=2&b=3"}
\usage{
list_to_string(filters)
}
\arguments{
\item{filters}{named list}
}
\value{
formatted string of parameters
}
\description{
Transform a list into a string of parameters
list("a" = list("1"), "b" = list("2", "3"))
becomes "a=1&b=2&b=3"
}
