% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_class.R
\name{remove_other_languages}
\alias{remove_other_languages}
\title{Remove columns that belong to other languages than the selected one}
\usage{
remove_other_languages(df, language)
}
\arguments{
\item{df}{data.frame returned from API with columns for all languages}

\item{language}{language to keep}
}
\value{
dataframe with columns relevant to selected language only
}
\description{
Remove columns that belong to other languages than the selected one
}
