% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_request.R
\name{sdmx_request}
\alias{sdmx_request}
\title{API query for SDMX output}
\usage{
sdmx_request(url)
}
\arguments{
\item{url}{url to query}
}
\value{
dataframe response
}
\description{
API query for SDMX output
}
